pyproject.toml
[tool.poetry]
name = "f4project"
version = "0.1.0"
description = ""
authors = ["gumba <gumba6740@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.14"
fastapi = "^0.135.1"
uvicorn = "^0.41.0"
pydantic = "^2.12.5"
asyncio = "^4.0.0"
tortoise-orm = "^0.21.0"
aerich = "^0.9.2"
gunicorn = "^25.1.0"
psycopg = {extras = ["binary"], version = "^3.3.3"}
asyncpg = "^0.31.0"
beautifulsoup4 = "^4.14.3"
requests = "^2.32.5"
lxml = "^6.0.2"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
dotenv = "^0.9.9"
pydantic-settings = "^2.13.1"
email-validator = "^2.3.0"


[tool.poetry.group.dev.dependencies]
black = "24.10.0"
ruff = "0.8.2"
mypy = "1.13.0"
pytest = "8.3.4"
coverage = "7.6.9"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.aerich]
tortoise_orm = "app.core.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.mypy]
python_version = "3.14"
strict = true
warn_unused_configs = true
warn_unused_ignores = true
